requirements.txt
ultralytics
opencv-python
pyserial

# python -m venv venv
# source venv/bin/activate    # Mac/Linux
# venv\Scripts\activate       # Windows
# pip install -r requirements.txt